fix(core): don't set legacy-peer-deps by default
Changing the value of `legacy-peer-deps` causes npm to install a
different graph, which can break builds and the `npm ci` command. Any
users who rely on `legacy-peer-deps` should already have it set in an
`.npmrc` file, so this should not affect them. It does, however, fix
bugs with using `nx` in repos that _don't_ have `legacy-peer-deps`
enabled, which is the vast majority of them.

Fixes #22066, fixes #29537