fix(core): handle nx config in package.json in move generator (#30270)
## Current Behavior

When the Nx configuration is in `package.json#nx` and not in
`project.json`, the move generator creates a `project.json` file in the
new destination.

## Expected Behavior

When the Nx configuration is in `package.json#nx` and not in
`project.json`, the move generator should not create a `project.json`
file in the new destination and should update the `nx` entry in the
`package.json` file.

## Related Issue(s)

Fixes #