docs(angular): fix typo in dynamic mfe section (#30311)
The documentation currently states:
> "Add a `module-federation.manifest.json` file to the `src/public/`
folder in our Dashboard application..."

However, the correct location for the file is `public/` and **not**
`src/public/`. This could lead to confusion when implementing the
example, as the folder structure may differ.

So I fixed it to public/

This issue was reported here: [Issue
#30020](https://github.com/nrwl/nx/issues/30020).

(cherry picked from commit d1a7ac96ceb04fc44199c059e97e388b9c492911)