docs(nx-dev): improve enterprise design & animations
In the nx-dev pages, adjust the interface to utilize 'EnterpriseLayout' instead of 'DefaultLayout'. The duration of the transition in 'customer-logos.tsx' has been reduced from 500 to 200 to boost responsiveness.

The 'enterprise-layout.tsx' file was added and contains new animation elements to enhance end-user experience. Updates were made to 'hero.tsx', simplifying and enhancing SVG animation.