docs(misc): document Vim/Nvim nx watch issue
If you are using Vim/Nvim and you are saving a file this file is not edited directly. The content is saved to a backup file and after this step the original file gets deleted and the backup file gets renamed. The nx watch command can't handle this at the moment.